Update index.md
Please can you pull in the updates to the QFAB (now part of QCIF) entry. I'd also like to submit a new logo, but am not able to push that to the src/images/logos folder, nor apparently attach it to this comment. How can I go about that please?